conformal_social_vae/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> -c pytorch
# platform: linux-64
#cudatoolkit=11.3.1=h2bc3f7f_2
#numpy=1.21.2
#python=3.9.7
#pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
#tensorboard=2.8.0=pyhd8ed1ab_1
#torchvision=0.12.0=py39_cu113

python=3.9
numpy=1.21
pytorch=1.11.0
torchvision=0.12.0
#cudatoolkit=11.3
tensorboard=2.8.0
